+++ packages/search-metadata-previews/src/snippet-preview/ProductDataDesktop.js
@@ -1,9 +1,10 @@
 import React, { Fragment } from "react";
 import PropTypes from "prop-types";
 import styled from "styled-components";
-import { __, sprintf } from "@wordpress/i18n";
+import { __, _n, sprintf } from "@wordpress/i18n";
 import { round, capitalize } from "lodash";
 import { StarRating } from "@yoast/components";
+import { DEFAULT_BEST_RATING } from "./constants";
 
 const ProductData = styled.span`
 	color: #70757a;
@@ -19,29 +20,32 @@ const ProductData = styled.span`
  */
 function ProductDataDesktop( props ) {
 	const { shoppingData } = props;
+	const { rating, bestRating = DEFAULT_BEST_RATING, reviewCount, availability, price } = shoppingData;
 
-	/* Translators: %s expands to the actual rating, e.g. 8/10. */
-	const ratingPart = sprintf( __( "Rating: %s", "wordpress-seo" ), round( ( shoppingData.rating * 2 ), 1 ) + "/10" );
+	/* Translators: %s expands to the actual rating, e.g. 4.5 or 4.5/6 */
+	const ratingPart = sprintf( __( "Rating: %s", "wordpress-seo" ), round( rating, 1 ) + ( bestRating === DEFAULT_BEST_RATING ? "" : `/${bestRating}` ) );
 
 	/* Translators: %s expands to the review count. */
-	const reviewPart = sprintf( __( "%s reviews", "wordpress-seo" ), shoppingData.reviewCount );
+	const reviewPart = sprintf( _n( "%s review", "%s reviews", reviewCount, "wordpress-seo" ), reviewCount );
+
+	const hasPriceOrAvailability = Boolean( price ) || Boolean( availability );
 
 	return (
 		<ProductData>
-			{ ( shoppingData.reviewCount > 0 ) &&
+			{ ( rating > 0 && rating <= bestRating && reviewCount > 0 ) &&
 				<Fragment>
-					<StarRating rating={ shoppingData.rating } />
+					<StarRating rating={ rating / bestRating * DEFAULT_BEST_RATING } />
 					<span> { ratingPart } · </span>
-					<span>{ reviewPart } · </span>
+					<span>{ reviewPart }{ hasPriceOrAvailability ? " · " : "" }</span>
 				</Fragment>
 			}
-			{ shoppingData.price &&
+			{ price &&
 				<Fragment>
-					<span dangerouslySetInnerHTML={ { __html: shoppingData.price } } />
+					<span dangerouslySetInnerHTML={ { __html: price } } />
 				</Fragment>
 			}
-			{ shoppingData.availability &&
-				<span>{ ` · ${ capitalize( shoppingData.availability ) }` }</span> }
+			{ availability &&
+				<span>{ ` · ${ capitalize( availability ) }` }</span> }
 		</ProductData>
 	);
 }